Refactor `local.py` into smaller logically organized files hosted under a `locally_connected` directory.
PiperOrigin-RevId: 424972748
