+++ keras/api/golden/v1/tensorflow.keras.layers.-locally-connected2-d.pbtxt
@@ -1,6 +1,6 @@
 path: "tensorflow.keras.layers.LocallyConnected2D"
 tf_class {
-  is_instance: "<class \'keras.layers.local.LocallyConnected2D\'>"
+  is_instance: "<class \'keras.layers.locally_connected.locally_connected2d.LocallyConnected2D\'>"
   is_instance: "<class \'keras.engine.base_layer.Layer\'>"
   is_instance: "<class \'tensorflow.python.module.module.Module\'>"
   is_instance: "<class \'tensorflow.python.training.tracking.autotrackable.AutoTrackable\'>"
