+++ keras/layers/serialization.py
@@ -29,7 +29,7 @@ from keras.layers import core
 from keras.layers import cudnn_recurrent
 from keras.layers import einsum_dense
 from keras.layers import embeddings
-from keras.layers import local
+from keras.layers import locally_connected
 from keras.layers import merging
 from keras.layers import noise
 from keras.layers import pooling
@@ -59,10 +59,10 @@ from tensorflow.python.util.tf_export import keras_export
 
 ALL_MODULES = (base_layer, input_layer, advanced_activations, attention,
                convolutional, convolutional_recurrent, core, cudnn_recurrent,
-               embeddings, einsum_dense, local, merging, batch_normalization_v1,
-               layer_normalization, unit_normalization, pooling,
-               image_preprocessing, recurrent, regularization, reshaping,
-               wrappers, hashing, hashed_crossing, category_encoding,
+               embeddings, einsum_dense, locally_connected, merging,
+               batch_normalization_v1, layer_normalization, unit_normalization,
+               pooling, image_preprocessing, recurrent, regularization,
+               reshaping, wrappers, hashing, hashed_crossing, category_encoding,
                discretization, integer_lookup, preprocessing_normalization,
                string_lookup, text_vectorization)
 ALL_V2_MODULES = (rnn_cell_wrapper_v2, batch_normalization, layer_normalization,
